#2225 :truck: move common component files in ergebnismeldung/common
