+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/common/BaseCardSnippedErgebnis.vue
