+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/common/TheErgebnisermittlungStimmzettelumschlaegeCard.vue
